abc184

+++ contests/atcoder/abc184/p3.cpp
@@ -4,22 +4,22 @@ using namespace std;
 typedef long long ll;
 #define pb push_back
 
-ll di[4] = {0, 3, 0, -3}, dj[4] = {3, 0, -3, 0};
+int di[4] = {0, 3, 0, -3}, dj[4] = {3, 0, -3, 0};
 
-bool check(ll r1, ll c1, ll r2, ll c2)
+bool check(int r1, int c1, int r2, int c2)
 {
-    for (ll i = -2; i <= 2; i++)
+    for (int i = -2; i <= 2; i++)
     {
-        for (ll j = -2; j <= 2; j++)
+        for (int j = -2; j <= 2; j++)
         {
-            ll ii = r1 + i, jj = c1 + j;
+            int ii = r1 + i, jj = c1 + j;
             if (abs(ii - r2) == abs(jj - c2))
                 return true;
         }
     }
     for (int dir = 0; dir < 4; dir++)
     {
-        ll ii = r1 + di[dir], jj = c1 + dj[dir];
+        int ii = r1 + di[dir], jj = c1 + dj[dir];
         if (abs(ii - r2) == abs(jj - c2))
             return true;
     }
@@ -28,7 +28,7 @@ bool check(ll r1, ll c1, ll r2, ll c2)
 
 int solve()
 {
-    ll r1, c1, r2, c2;
+    int r1, c1, r2, c2;
     cin >> r1 >> c1 >> r2 >> c2;
     if (r1 == r2 && c1 == c2)
         return 0;
